Initial Client and Server part of Goof Talk
Instructions to execute
go build both server and client separately.
give the executable permission to the object code.
run server first in a terminal
then open n terminals to register n users and type the command in the
below format
./client -user username -host localhost

this shows n connection in the same system.
to  have this application running on different pc change the ip address.